Fix `MeasurementController.rename_value` for multiple columns
Partial fix for #187. Updating the measurement table changing a conditionId in the condition table also failed for me, although with a different error than in #187.
That was related to using unsupported multidimensional indexers which is addressed here.

Visualization after rename will fail due to invalid condition IDs. To be fixed elsewhere. I guess, if a condition is renamed, not only the meaurement table should be updated, but also the visualization table.